requirements.txt
Django==5.0.1
djangorestframework
django-cors-headers
djangorestframework-simplejwt # Implémentation du jeton JWT pour DRF 
dj-rest-auth    # Utilitaires d'authentification REST API
mysqlclient
dj-rest-auth[with_social]
django-allauth # Pour la gestion des comptes utilisateurs et l'authentification sociale
django-phonenumber-field[phonenumbers]
pillow
python-decouple
django-filter
drf-yasg
django-jazzmin
psycopg2-binary # PostgreSQL database adapter for Python
python-dotenv # Pour la gestion des variables d'environnement
psycopg2-2.9.10